% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_comb.R
\name{all_comb}
\alias{all_comb}
\title{Analyze all combinations}
\usage{
all_comb(gdist.inputs, 
                GA.inputs, 
                results.dir,
                max.combination = 4,
                iters = 1000,
                replicate = 1,
                sample.prop = 0.75,
                nlm = FALSE,
                dist_mod = TRUE,
                null_mod = TRUE)
}
\arguments{
\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function. Be sure that the \code{Results.dir} has been been correctly specified as "all.comb"}

\item{results.dir}{Directory to write and save analysis results. This should be an empty directory. Any existing files located in this directory will be deleted!}

\item{max.combination}{The maximum number of surfaces to include in the all combinations analysis (Default = 4). Alternatively, specify a vector with the minimum and maximum number of surfaces to combine (e.g., c(2,4). If the minimum > 1, then the single surface optimization will be skipped.}

\item{iters}{Number of bootstrap iterations to be conducted (Default = 1000)}

\item{replicate}{The number of times to replicate the GA optimization process for each surface (Default = 1)}

\item{sample.prop}{Proportion of observations to be sampled each iteration (Default = 0.75)}

\item{nlm}{Logical, if TRUE, the final step of optimization will use nlm to fine-tune parameter estimates. This may lead to overfitting in some cases. (Default = FALSE)}

\item{dist_mod}{Logical, if TRUE, a Distance model will be calculated and added to the output table (default = TRUE)}

\item{null_mod}{Logical, if TRUE, an intercept-only model will be calculated and added to the output table (Default = TRUE)}
}
\value{
This function optimizes resistance surfaces in isolation using \code{\link[ResistanceGA]{SS_optim}}, followed by multisurface optimization using \code{\link[ResistanceGA]{MS_optim}}, and then conducts a bootstrap analysis.
}
\description{
A wrapper function to run both \code{\link[ResistanceGA]{SS_optim}} and \code{\link[ResistanceGA]{MS_optim}} to optimize all combinations of resistance surfaces with the Genetic Algorithms package \pkg{GA}. Following optimization, \code{\link[ResistanceGA]{Resist.boot}} is run to conduct a bootstrap analysis.  This function can only be used when optimizing resistance surface with least cost path or commute distance (\pkg{gdistance}).
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
