% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{Dc_list}
\alias{Dc_list}
\title{Data file. A list of three matrices}
\format{
A list of length 3
}
\usage{
data(Dc_list)
}
\description{
Sample file to be used with examples in the vignette
}
\details{
Each matrix depicts the pairwise genetic distance (measured as chord distance) between sample locations

 \itemize{    
   \item Dc_cat. Pairwise chord distance matrix simulated across the categorical resistance surface
   \item Dc_cont. Pairwise chord distance matrix simulated across the continuous resistance surface
   \item Dc_multi. Pairwise chord distance matrix simulated across the multivariate resistance surface
   }
}
\keyword{datasets}
