% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_Search.R
\name{Grid.Search}
\alias{Grid.Search}
\title{Conduct grid search of response surface}
\usage{
Grid.Search(shape, max, transformation, Resistance, CS.inputs, gdist.inputs, GA.inputs)
}
\arguments{
\item{shape}{A vector of values for the shape parameter}

\item{max}{A vector of values for the maximum value parameter}

\item{transformation}{Transformation to apply. Can be either numeric or character of transformation name}

\item{Resistance}{An R Raster object, or path to a .asc file}

\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function}
}
\value{
This function will return a \code{filled.contour} plot. Additionally, an object with values that can be plotted with \code{filled.contour} to visualize the response surface
}
\description{
Visualize the AICc response surface
}
\details{
This function will perform a full factorial grid search of the values provided in the shape and max.scale vectors. Depending on the number of values provided for each, and the time it takes to run each iteration, this process may take a while to complete. \cr Suitable values for transformation:\cr
\tabular{ll}{
   \tab 1 = "Inverse-Reverse Monomolecular"\cr
   \tab 2 = "Inverse-Reverse Ricker"\cr
   \tab 3 = "Monomolecular"\cr
   \tab 4 = "Ricker"\cr
   \tab 5 = "Reverse Monomolecular"\cr
   \tab 6 = "Reverse Ricker"\cr
   \tab 7 = "Inverse Monomolecular"\cr
   \tab 8 = "Inverse Ricker"\cr
   \tab 9 = "Distance"\cr
   }
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
