% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{jl_out}
\alias{jl_out}
\title{Data file. Example raster output from CIRCUITSCAPE Julia}
\format{
A RasterLayer
}
\usage{
data(jl_out)
}
\description{
Sample file used with examples in the vignette
}
\details{
A raster file
}
\keyword{datasets}
