% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_smooth.R
\name{k.smooth}
\alias{k.smooth}
\title{Kernel smoothing function}
\usage{
k.smooth (raster, sigma)
}
\arguments{
\item{raster}{A RasterLayer object to be smoothed}

\item{sigma}{The standard deviation of the Gaussian smoothing parameter (see \code{\link[smoothie]{kernel2dsmooth}} documentation in the \code{smoothie} package.)}
}
\description{
Apply Gaussian kernel smoothing of specified sigma
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
