% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MLPE.lmm_coef}
\alias{MLPE.lmm_coef}
\title{Obtain coefficients from maximum likelihood population effects mixed effects model (MLPE)}
\usage{
MLPE.lmm_coef(resist.dir, genetic.dist, out.dir)
}
\arguments{
\item{resist.dir}{Directory containing optimized .asc resistance files (optimized resistance surfaces)}

\item{genetic.dist}{Lower half of pairwise genetic distance matrix}

\item{out.dir}{If specified, a .csv table will printed to the specified directory (Default = NULL)}
}
\value{
A table of MLPE fitted model coefficients
}
\description{
Runs MLPE as detailed by Clarke et al. (2002). This function is designed to generate a table of the fitted mixed effects model coefficients
}
\references{
Clarke, R. T., P. Rothery, and A. F. Raybould. 2002. Confidence limits for regression relationships between distance matrices: Estimating gene flow with distance. Journal of Agricultural, Biological, and Environmental Statistics 7:361-372.
}

