% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MS_optim}
\alias{MS_optim}
\title{Simultaneous optimization of multiple resistance surfaces}
\usage{
MS_optim(CS.inputs, gdist.input, GA.inputs)
}
\arguments{
\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function}
}
\value{
This function optimizes multiple resistance surfaces, returning a Genetic Algorithm (GA) object with summary information. Diagnostic plots of model fit are output to the "Results/Plots" folder that is automatically generated within the folder containing the optimized ASCII files. A text summary of the optimization settings and results is printed to the results folder.
}
\description{
Optimize multiple resistance surfaces simultaneously using genetic algorithms
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

