% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{raster_orig}
\alias{raster_orig}
\title{Data file. Simulated original resistance surfaces}
\format{RasterStack object of length 2}
\usage{
data(raster_orig)
}
\description{
A raster stack containing two raster surfaces generated from the \code{RandomFields} package. These surfaces are the raw, original surfaces (prior to transformation) used with individual-based genetic simulations conducted using \code{PopGenReport}. These are the inputs for optimization examples.
}
\details{
\itemize{    
   \item cat_orig. 5-class categorical resistance surface with equal representation of each categorcal level.
   \item cont_orig. Continuous resistance surface.
   }
}
\keyword{datasets}
