% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{raster_true}
\alias{raster_true}
\title{Data file. Simulated true resistance surfaces}
\format{
RasterStack object of length 3
}
\usage{
data(raster_true)
}
\description{
A raster stack containing three raster surfaces generated from the \code{RandomFields} package. These surfaces were used with individual-based genetic simulations conducted using \code{PopGenReport}.
}
\details{
\itemize{    
   \item cat_true. 4-class categorical resistance surface. Derived from reclass: 1 = 1, 2 = 75, 3 = 200, 4 = 25, 5 = 1.
   \item cont_true. Continuous resistance surface. Derived from inverse monomolecular transformation with a shape parameter of 2.75, and maximum resistance parameter of 100.
   \item multi_true. A composite surface created by summing transformed categorical and continuous surfaces, rescaled to a minimum value of 1.
   }
}
\keyword{datasets}
