% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{resist_list}
\alias{resist_list}
\title{Data file. A list of three matrices}
\format{
A list of length 3
}
\usage{
data(resist_list)
}
\description{
Sample file to be used with examples in the vignette
}
\details{
Each matrix depicts the pairwise effective resistance distance (measured using `gdistance`) between sample locations across `resist_true` surfaces

 \itemize{      
   \item resist_cat. Pairwise effective distance matrix calculated across the categorical resistance surface
   \item resist_cont. Pairwise effective distance matrix calculated across the continuous resistance surface
   \item resist_cont. Pairwise effective distance matrix calculated across the multivariate resistance surface
   }
}
\keyword{datasets}
