% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Resistance.Opt_multi}
\alias{Resistance.Opt_multi}
\title{Optimize multiple resistance surfaces simultaneously}
\usage{
Resistance.Opt_multi(PARM, CS.inputs = NULL, gdist.inputs = NULL, GA.inputs,
  Min.Max, quiet = FALSE)
}
\arguments{
\item{PARM}{Parameters to transform conintuous surface or resistance values of categorical surface. Should be a vector with parameters specified in the order of resistance surfaces. These values are selected during optimization if called within GA function.}

\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function}

\item{Min.Max}{Define whether the optimization function should minimized ('min') or maximized ('max')}

\item{quiet}{Logical, if TRUE, AICc and iteration time will not be printed to the screen at the completion of each iteration. Default = FALSE}
}
\value{
AIC value from mixed effect model
}
\description{
Create composite resistance surface by simultaneously optimizing multiple categoricla and continuous surfaces. This optimization function is designed to be called from GA
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

