% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimFunction_Multi_scale.R
\name{Resistance.Opt_multi.scale}
\alias{Resistance.Opt_multi.scale}
\title{Optimize multiple resistance surfaces simultaneously with kernel smoothing}
\usage{
Resistance.Opt_multi.scale(PARM, CS.inputs = NULL, gdist.inputs = NULL,
  jl.inputs = NULL, GA.inputs, Min.Max, quiet = FALSE)
}
\arguments{
\item{PARM}{Parameters to transform conintuous surface or resistance values of categorical surface. Should be a vector with parameters specified in the order of resistance surfaces. These values are selected during optimization if called within GA function.}

\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{jl.inputs}{Object created from running \code{\link[ResistanceGA]{jl.prep}} function. Defined if optimizing using CIRCUITSCAPE run in Julia}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function}

\item{Min.Max}{Define whether the optimization function should minimized ('min') or maximized ('max')}

\item{quiet}{Logical, if TRUE, AICc and iteration time will not be printed to the screen at the completion of each iteration. Default = FALSE}
}
\value{
Objective function value (either AIC, R2, or LL) from mixed effect model
}
\description{
Create composite resistance surface by simultaneously optimizing multiple binary and continuous surfaces to optimized with a kernel smoothing parameter. This optimization function is designed to be called from GA
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
