% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimFunction_Single_covariate.R
\name{Resistance.Opt_single.cov}
\alias{Resistance.Opt_single.cov}
\title{Optimize resistance surfaces individually with more than one predictor variable}
\usage{
Resistance.Opt_single.cov(PARM, Resistance, CS.inputs = NULL,
  gdist.inputs = NULL, jl.inputs = NULL, GA.inputs, Min.Max = "max",
  iter = NULL, quiet = FALSE)
}
\arguments{
\item{PARM}{Parameters to transform conintuous surface or resistance values of categorical surface. A vector with parameters specified in the order of resistance surfaces.These values are selected during optimization if called within GA function.}

\item{Resistance}{Resistance surface to be optimized. This should be an R raster object. If not specified, the function will attempt to find the a resistance surface from \code{GA.inputs}}

\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{jl.inputs}{Object created from running \code{\link[ResistanceGA]{jl.prep}} function. Defined if optimizing using CIRCUITSCAPE run in Julia}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function}

\item{Min.Max}{Define whether the optimization function should minimized ('min') or maximized ('max'). Default in 'max'}

\item{iter}{A counter for the number of surfaces that will be optimized}

\item{quiet}{Logical, if TRUE AICc and iteration duration will not be printed to the screen at the completion of each iteration.}
}
\value{
Objective function value (either AIC, R2, or LL) from mixed effect model
}
\description{
Optimize all resistance surfaces that are located in the same directory individually. This optimization function is designed to be called from GA
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
