% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{resistance_surfaces}
\alias{resistance_surfaces}
\title{Raster stack of resistance surfaces}
\format{Raster stack object of length 4}
\usage{
resistance_surfaces
}
\description{
There are four raster surfaces contained in the \code{resistance_surfaces} raster stack: two continuous surfaces, one 3-class categorical surface, and one 2-class feature surface
}
\keyword{datasets}

