% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{response}
\alias{response}
\title{Example pairwise genetic response data}
\format{A 25 x 25 matrix}
\usage{
data(response)
}
\description{
Sample file that can be used with examples in the vignette. These simulated pairwise resistance distances represent the 'true' response to a composite resistance surface consisting of an inverse Ricker transformation of the the example continuous surface, and a reclassification of the categorical surface such that 0 = 1, 1 = 250, and 2 = 75.
}
\details{
A square matrix of simulated pairwise distances
}
\keyword{datasets}
