% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_julia.R
\name{Run_CS.jl}
\alias{Run_CS.jl}
\title{Run Julia version of CIRCUITSCAPE from R}
\usage{
Run_CS.jl(CS.inputs, 
r, 
CurrentMap = FALSE, 
full.mat = FALSE,
EXPORT.dir = NULL, 
output = "matrix")
}
\arguments{
\item{jl.inputs}{Object created from running \code{\link[ResistanceGA]{jl.prep}} function}

\item{r}{Accepts two types of inputs. Provide either the path to the resistance surface file (.asc) or specify an R RasterLayer object}

\item{CurrentMap}{Logical. If TRUE, the cumulative current resistance map will be generated during the CS run (Default = FALSE)}

\item{full.mat}{Logical (Default = FALSE). If TRUE, the full distance matrix will be generated as an R object, rather than just the lower half of the distance matrix.}

\item{EXPORT.dir}{Directory where CS results will be written. Only specify if CIrcuitscape current map outputs are requested. It is critical that there are NO SPACES in the directory, as this may cause the function to fail.}

\item{output}{Specifiy either "matrix" or "raster". "matrix" will return the lower half of the pairwise resistance matrix (default), while "raster" will return a \code{raster} object of the cumulative current map. The raster map can only be returned if \code{CurrentMap=TRUE}}
}
\value{
Vector of CIRCUITSCAPE resistance distances (lower half of "XXX_resistances.out") OR a full square distance matrix if `full.mat` = TRUE. Alternatively, a raster object of the cumulative current map can be returned when \code{CurrentMap=TRUE} and \code{output="raster"}.
}
\description{
Execute julia CS from R
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
