% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHNe.R
\name{SHNe}
\alias{SHNe}
\title{Calculate distance-based Spatial-Heterogeneity-in-Ne (SHNe) metrics}
\usage{
SHNe(n.samples,
            pop.size,
            out = 'df')
}
\arguments{
\item{n.samples}{Specify the number of samples (populations) included in your analysis}

\item{pop.size}{Provide a vector of values that indicates Ne for each sampled population. These values should be ordered the same as your sample populations}

\item{output}{Default is 'df', will return a data frame with distance measures. If 'matrix' is specified, then a list containing a matrix for each distance measure is produced.}
}
\value{
Either a data frame or a list of matrices containing the SHNe metrics
}
\description{
Use formulas from Prunier et al. to calculate di and dhm
}
\references{
Prunier, J. G., V. Dubut, L. Chikhi, and S. Blanchet. 2017. Contribution of spatial heterogeneity in effective population sizes to the variance in pairwise measures of genetic differentiation. Methods in Ecology and Evolution 8:1866-1877.

#' @examples  
## Create vector of Ne
Ne <- rep(c(5, 25, 100), each = 3)
SHNe.out <- SHNe(n.samples = length(Ne),
                 pop.size = Ne
                 )
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>

Bill Peterman <Bill.Peterman@gmail.com>
}
