% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_optim.R
\name{SS_optim}
\alias{SS_optim}
\title{Single surface optimization}
\usage{
SS_optim(CS.inputs, gdist.inputs, GA.inputs, nlm, dist_mod, null_mod)
}
\arguments{
\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function}

\item{nlm}{Logical, if TRUE, the final step of optimization will use nlm to fine-tune parameter estimates. This may lead to overfitting in some cases. Default = FALSE.}

\item{dist_mod}{Logical, if TRUE, a Distance model will be calculated and added to the output table (default = TRUE)}

\item{null_mod}{Logical, if TRUE, an intercept-only model will be calculated and added to the output table (default = TRUE)}
}
\value{
This function optimizes resistance surfaces in isolation. Following optimization of all surfaces, several summary objects are created.\cr
\enumerate{
\item Diagnostic plots of model fit are output to the "Results/Plots" directory that is automatically generated within the folder containing the optimized ASCII files.
\item A .csv file with the Maximum Likelihood Population Effects mixed effects model coefficient estimates (MLPE_coeff_Table.csv)
\item Three summary .csv files are generated: CategoricalResults.csv, ContinuousResults.csv, & All_Results_AICc.csv. These tables contain AICc values and optimization summaries for each surface.
}
All results tables are also summarized in a named list ($ContinuousResults, $CategoricalResults, $AICc, $MLPE, $MLPE.list, $cd, $k)\cr
The \code{lmer} model objects stored $MLPE.list are fit using Restricted Maximum Likelihood \cr
$cd is a list of the optimized cost pairwise distance matrices and $k is a table of the surface names and number of parameters used to calculate AICc. These two objects can be passed to \code{\link[ResistanceGA]{Resist.boot}} to conduct a bootstrap analysis.
}
\description{
Optimize all surfaces contained in a directory using a genetic algorithm executed with the \code{\link[GA]{ga}} function in the Genetic Algorithms package \pkg{GA}
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
