% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_optim_scale.R
\name{SS_optim.scale}
\alias{SS_optim.scale}
\title{Single surface optimization with kernel smoothing}
\usage{
SS_optim.scale(CS.inputs, 
  gdist.inputs, 
  jl.inputs,
  GA.inputs,
  dist_mod, 
  null_mod)
}
\arguments{
\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{jl.inputs}{Object created from running \code{\link[ResistanceGA]{jl.prep}} function. Defined if optimizing using CIRCUITSCAPE run in Julia}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function}

\item{dist_mod}{Logical, if TRUE, a Distance model will be calculated and added to the output table (default = TRUE)}

\item{null_mod}{Logical, if TRUE, an intercept-only model will be calculated and added to the output table (default = TRUE)}
}
\value{
This function optimizes resistance surfaces in isolation. Following optimization of all surfaces, several summary objects are created.\cr
\enumerate{
\item Diagnostic plots of model fit are output to the "Results/Plots" directory that is automatically generated within the folder containing the optimized ASCII files.
\item A .csv file with the Maximum Likelihood Population Effects mixed effects model coefficient estimates (MLPE_coeff_Table.csv)
\item Three summary .csv files are generated: CategoricalResults.csv, ContinuousResults.csv, & All_Results_AICc.csv. These tables contain AICc values and optimization summaries for each surface.
}
All results tables are also summarized in a named list ($ContinuousResults, $CategoricalResults, $AICc, $MLPE, $MLPE.list)\cr
The \code{lmer} model objects stored $MLPE.list are fit using Restricted Maximum Likelihood
}
\description{
Optimize all binary and/or continuous surfaces contained in a directory using a genetic algorithm executed with the \code{\link[GA]{ga}} function in the Genetic Algorithms package \pkg{GA}. Optimizes a kernel smoothing parameter with all surfaces.
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
