% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResistanceGA_Functions.R
\name{Combine_Surfaces}
\alias{Combine_Surfaces}
\title{Combine multiple resistance surfaces together}
\usage{
Combine_Surfaces(PARM, CS.inputs, gdist.inputs, GA.inputs, out, File.name, rescale)
}
\arguments{
\item{PARM}{Parameters to transform conintuous surface or resistance values of categorical surface. Requires a vector with parameters specified in the order of resistance surfaces}

\item{CS.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function. Defined if optimizing using CIRCUITSCAPE}

\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA]{GA.prep}} function.}

\item{out}{Directory to write combined .asc file. Default = NULL and no files are exported}

\item{File.name}{Name of output .asc file. Default is the combination of all surfaces combined, separated by "."}

\item{rescale}{Locical. If TRUE (default), the values of the combined raster surface will be divided by the minimum value to create a resistance surface with a minimum value = 1.}
}
\value{
R raster object that is the sum all transformed and/or reclassified resistance surfaces provided
}
\description{
Combine multiple resistance surfaces into new composite surface based on specified parameters
}
\details{
\code{PARM} is designed to accept the output of \code{MS_optim}. For continuous surfaces, there are three terms: 1) Transformation, 2) shape, and 3) maximum value. Transformation must be provided as a numeric value:\cr
\tabular{ll}{
   \tab 1 = "Inverse-Reverse Monomolecular"\cr
   \tab 2 = "Inverse-Reverse Ricker"\cr
   \tab 3 = "Monomolecular"\cr
   \tab 4 = "Ricker"\cr
   \tab 5 = "Reverse Monomolecular"\cr
   \tab 6 = "Reverse Ricker"\cr
   \tab 7 = "Inverse Monomolecular"\cr
   \tab 8 = "Inverse Ricker"\cr
   \tab 9 = "Distance"\cr
   }

The Distance transformation sets all values equal to one. Because of the flexibility of the Ricker function to take a monomolecular shape (try \code{Plot.trans(PARM=c(10,100), Resistance=c(1,10), transformation="Ricker")} to see this), whenever a shape parameter >6 is selected in combination with a Ricker family transformation, the transformation reverts to a Distance transformation. In general, it seems that using a combination of intermediate Ricker and Monomolecular transformations provides the best, most flexible coverasge of parameter space.
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

