% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResistanceGA_Functions.R
\name{CS.prep}
\alias{CS.prep}
\title{Prepare and bundle input CIRCUITSCAPE model parameters}
\usage{
CS.prep(n.POPS, response, CS_Point.File, CS.program, Neighbor.Connect, pairs_to_include)
}
\arguments{
\item{n.POPS}{The number of populations that are being assessed}

\item{response}{Vector of pairwise genetic distances (lower half of pairwise matrix).}

\item{CS_Point.File}{The path to the Circuitscape formatted point file. See Circuitscape documentation for help.}

\item{CS.program}{The path to the CIRCUITSCAPE executable file (cs_run.exe) on a Windows PC. See details below.}

\item{Neighbor.Connect}{Select 4 or 8 to designate the connection scheme to use in CIRCUITSCAPE (Default = 8)}

\item{pairs_to_include}{Default is NULL. If you wish to use the advanced CIRCUITSCAPE setting mode to include or exclude certain pairs of sample locations, provide the path to the properly formatted "pairs_to_include.txt" file here. Currently only "include" method is supported.}
}
\value{
An R object that is a required input into optimization functions
}
\description{
This function will prepare objects needed for running optimization functions
}
\details{
\code{CS.program} Example of path to CIRCUITSCAPE executible on Windows:

'"C:/Program Files/Circuitscape/cs_run.exe"'

***NOTE: Double quotation used***
This is the current default for \code{CS.program}, but the directory may need to be changed depending upon your installation of CIRCUITSCAPE
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

