% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CS.prep}
\alias{CS.prep}
\title{Prepare and bundle input CIRCUITSCAPE model parameters}
\usage{
CS.prep(n.POPS, response, CS_Point.File, CS.program, Neighbor.Connect, platform)
}
\arguments{
\item{n.POPS}{The number of populations that are being assessed}

\item{response}{Vector of pairwise genetic distances (lower half of pairwise matrix).}

\item{CS_Point.File}{The path to the Circuitscape formatted point file. See Circuitscape documentation for help.}

\item{CS.program}{The path to the CIRCUITSCAPE executable file (cs_run.exe) if using a Windows PC. If using Linux or Mac, this should be the directory to the CIRCUITSCAPE python file (cs_run.py). See details below.}

\item{Neighbor.Connect}{Select 4 or 8 to designate the connection scheme to use in CIRCUITSCAPE (Default = 8)}

\item{platform}{What computing platform are you using ("pc", "other")}
}
\value{
An R object that is a required input into optimization functions
}
\description{
This function will prepare objects needed for running optimization functions
}
\details{
\code{CS.exe} Example of path to CIRCUITSCAPE executible on Windows:

'"C:/Program Files/Circuitscape/cs_run.exe"'

***NOTE: Double quotation used***
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

