% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Diagnostic.Plots}
\alias{Diagnostic.Plots}
\title{Create diagnostic plots}
\usage{
Diagnostic.Plots(resistance.mat, genetic.dist, XLAB,YLAB, plot.dir, type, name)
}
\arguments{
\item{resistance.mat}{Path to CIRCUITSCAPE "_resistances.out" file, or costDistance object created from running gdistance}

\item{genetic.dist}{Vector of pairwise genetic distances (lower half of pairwise matrix). Can be input as CS.inputs$response}

\item{XLAB}{Label for x-axis (Defaults to "Estimated resistance")}

\item{YLAB}{Label for y-axis (Defaults to "Genetic distance")}

\item{plot.dir}{Directory to output TIFF of diagnostic plots}

\item{type}{Specify whether the optimized surface is "continuous" or "categorical"}

\item{name}{The name to be attached to the output file. Must be specified when getting diagnostic plots for gdistance models}
}
\value{
A multipanel panel .tif including histogram of residuals and qqplot of fitted mixed effects model
}
\description{
This function will generate mixed effect model diagnostic plots following optimization
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

