% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic_Plots.R
\name{Diagnostic.Plots}
\alias{Diagnostic.Plots}
\title{Create diagnostic plots}
\usage{
Diagnostic.Plots(resistance.mat, genetic.dist, XLAB,YLAB, plot.dir, type, name, ID, ZZ)
}
\arguments{
\item{resistance.mat}{Path to CIRCUITSCAPE "_resistances.out" file, or costDistance object created from running gdistance}

\item{genetic.dist}{Vector of pairwise genetic distances (lower half of pairwise matrix). Can be input as CS.inputs$response}

\item{XLAB}{Label for x-axis (Defaults to "Estimated resistance")}

\item{YLAB}{Label for y-axis (Defaults to "Genetic distance")}

\item{plot.dir}{Directory to output TIFF of diagnostic plots}

\item{type}{Specify whether the optimized surface is "continuous" or "categorical"}

\item{name}{The name to be attached to the output file. Must be specified when getting diagnostic plots for gdistance models}

\item{ID}{The to_from ID list for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of CS.prep or gdist.prep (Default = NULL)}

\item{ZZ}{The sparse matrix object for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of CS.prep or gdist.prep (Default = NULL)}
}
\value{
A multipanel panel .tif including histogram of residuals and qqplot of fitted mixed effects model
}
\description{
This function will generate mixed effect model diagnostic plots following optimization
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
