% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_prep.R
\name{GA.prep}
\alias{GA.prep}
\title{Create R object with genetic algorithm optimization settings}
\usage{
GA.prep(ASCII.dir,
               Results.dir = NULL,
               min.cat = NULL,
               max.cat = 2500,
               max.cont = 2500,
               min.scale = NULL,
               max.scale = NULL,
               shape.min = NULL,
               shape.max = NULL,
               cont.shape = NULL,
               select.trans = NULL,
               method = "LL",
               scale = FALSE,
               scale.surfaces = NULL,
               k.value = 2,
               pop.mult = 15,
               percent.elite = 0.05,
               type = "real-valued",
               pcrossover = 0.85,
               pmutation = 0.125,
               maxiter = 1000,
               run = NULL,
               keepBest = TRUE,
               population = gaControl(type)$population,
               selection = gaControl(type)$selection,
               crossover = "gareal_blxCrossover",
               mutation = gaControl(type)$mutation,
               pop.size = NULL,
               parallel = FALSE,
               gaisl = FALSE,
               island.pop = 20,
               numIslands = NULL,
               migrationRate = NULL,
               migrationInterval = NULL,
               optim = FALSE,
               optim.method = "L-BFGS-B", 
               poptim = 0.0,
               pressel = 1.00,
               control = list(fnscale = -1, maxit = 100),
               hessian = FALSE,
               opt.digits = NULL,
               seed = NULL,
               monitor = TRUE,
               quiet = FALSE)
}
\arguments{
\item{ASCII.dir}{Directory containing all raster objects to optimized. If optimizing using least cost paths, a RasterStack or RasterLayer object can be specified.}

\item{Results.dir}{If a RasterStack is provided in place of a directory containing .asc files for ASCII.dir, then a directory to export optimization results must be specified. It is critical that there are NO SPACES in the directory, as this will cause the function to fail. If using the \code{\link[ResistanceGA]{all_comb}} function, specify \code{Results.dir} as "all_comb".}

\item{min.cat}{The minimum value to be assessed during optimization of categorical resistance surfaces (Default = 1 / max.cat)}

\item{max.cat}{The maximum value to be assessed during optimization of categorical resistance surfaces (Default = 2500)}

\item{max.cont}{The maximum value to be assessed during optimization of continuous resistance surfaces (Default = 2500)}

\item{min.scale}{The minimum scaling parameter value to be assessed during optimization of resistance surfaces with kernel smoothing (Default = 0.01). See details}

\item{max.scale}{The maximum scaling parameter value to be assessed during optimization of resistance surfaces with kernel smoothing (Default = 0.1 * maximum dimension of the raster surface)}

\item{shape.min}{The minimum value for the shape parameter used for transformimg resistance surfaces. If unspecified, used 0.5}

\item{shape.max}{The maximum value for the shape parameter used for transformimg resistance surfaces. If unspecified, used 14.5}

\item{cont.shape}{A vector of hypothesized relationships that each continuous resistance surface will have in relation to the genetic distance response (Default = NULL; see details)}

\item{select.trans}{Option to specify which transformations are applied to continuous surfaces. Must be provided as a list. "A" = All, "M" = Monomolecular only, "R" = Ricker only. Default = "M"; see Details.}

\item{method}{Objective function to be optimized. Select "AIC", "R2", or "LL" to optimize resistance surfaces based on AIC, variance explained (R2), or log-likelihood. (Default = "LL")}

\item{scale}{Logical. To optimize a kernel smoothing scaling parameter during optimization, set to TRUE (Default = FALSE). See Details below.}

\item{scale.surfaces}{(Optional) If doing multisurface optimization with kernel smoothing, indicate which surfaces should be smoothed. A vector equal in length to the number of resistance surfaces to be optimized using MS_optim.scale that is used to indicate whether a surface should (1) or should not (0) have kernel smoothing applied. See details.}

\item{k.value}{Specification of how k, the number of parameters in the mixed effects model, is determined. Specify 1, 2, 3, or 4 (Default = 2; see details).

1 --> k = 2;

2 --> k = number of parameters optimized plus the intercept;

3 --> k =  the number of parameters optimized plus the intercept and the number of layers optimized;

4 --> k = the number of layers optimized plus the intercept}

\item{pop.mult}{Value will be multiplied with number of parameters in surface to determine 'popSize' in GA. By default this is set to 15.}

\item{percent.elite}{Percent used to determine the number of best fitness individuals to survive at each generation ('elitism' in GA). By default the top 5\% individuals will survive at each iteration.}

\item{type}{Default is "real-valued"}

\item{pcrossover}{Probability of crossover. Default = 0.85}

\item{pmutation}{Probability of mutation. Default = 0.125}

\item{maxiter}{Maximum number of iterations to run before the GA search is halted. If using standard \code{ga} optimizer, the default = 1000. If using \code{gaisl = TRUE}, then this is set to 15x the \code{migrationInterval}}

\item{run}{Number of consecutive generations or epochs without any improvement in objective function before the GA is stopped. If using standard \code{ga}, the default = 25. If using \code{gaisl = TRUE}, then the default \code{run} value will be calculated as \code{migrationInterval} * 5}

\item{keepBest}{A logical argument specifying if best solutions at each iteration should be saved (Default = TRUE)}

\item{population}{Default is "gareal_Population" from GA}

\item{selection}{Default is "gareal_lsSelection" from GA}

\item{crossover}{Default = "gareal_blxCrossover". This crossover method greatly improved optimization during preliminary testing}

\item{mutation}{Default is "gareal_raMutation" from GA}

\item{pop.size}{Number of individuals to create each generation. If \code{gaisl = TRUE}, then this number is automatically calculated as \code{numIslands} * \code{island.pop}}

\item{parallel}{A logical argument specifying if parallel computing should be used (TRUE) or not (FALSE, default) for evaluating the fitness function. You can also specify the number of cores to use. Parallel processing currently only works when optimizing using least cost paths. It will fail if used with CIRCUITSCAPE, so this is currently not an option.}

\item{gaisl}{Should the genetic algorithm use the islands parallel optimization? (Default = FALSE)}

\item{island.pop}{The number of individuals to populate each island. (Default = 20)}

\item{numIslands}{If \code{gaisl = TRUE}, an integer value which specifies the number of islands to use in the genetic evolution (by default will be set to 4)}

\item{migrationRate}{If \code{gaisl = TRUE}, a value in the range (0, 1) which gives the proportion of individuals that undergo migration between islands in every exchange (by default equal to 0.10).}

\item{migrationInterval}{If \code{gaisl = TRUE}, an integer value specifying the number of iterations at which exchange of individuals takes place. This interval between migrations is called an epoch, and it is set at 10 by default.}

\item{optim}{A logical defaulting to \code{FALSE} determining whether or not a local search using general-purpose optimisation algorithms should be used. See argument \code{optimArgs} for further details and finer control. Setting to TRUE has the potential to improve optimization accuracy, but will increase optimization time.}

\item{optim.method}{The method to be used among those available in \code{\link[stats]{optim}} function. By default, the BFGS algorithm with box constraints is used, where the bounds are those provided in the \code{ga()} function call. Further methods are available as described in the Details section in help(optim).}

\item{poptim}{A value in the range [0,1] specifying the probability of performing a local search at each iteration of GA (default 0.0). Only change if your optimization is relatively fast.}

\item{pressel}{A value in the range [0,1] specifying the pressure selection (default 1.00). The local search is started from a random solution selected with probability proportional to fitness. High values of pressel tend to select the solutions with the largest fitness, whereas low values of pressel assign quasi-uniform probabilities to any solution.}

\item{control}{A list of control parameters. See 'Details' section in \code{\link[stats]{optim}}}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned? This will allow for the calculation of standard errors on paramter estimates (not yet implemented). Default = FALSE}

\item{opt.digits}{The number of significant digits that the objective function will be assessed at. By default, no rounding occurs.}

\item{seed}{Integer random number seed to replicate \code{ga} optimization}

\item{monitor}{Default = TRUE, which prints the average and best fitness values at each iteration.}

\item{quiet}{Logical. If TRUE, the objective function and step run time will not be printed to the screen after each step. Only \code{ga} summary information will be printed following each iteration. (Default = FALSE)}
}
\value{
An R object that is a required input into optimization functions
}
\description{
This function prepares and compiles objects and commands for optimization with the GA package
}
\details{
Only files that you wish to optimize, either in isolation or simultaneously, should be included in the specified \code{ASCII.dir}. If you wish to optimize different combinations of surfaces, different directories containing these surfaces must be created.

When \code{scale = TRUE}, the standard deviation of the Gaussian kernel smoothing function (sigma) will also be optimized during optimization. Only continuous surfaces or binary categorical surfaces (e.g., forest/no forest; 1/0) surfaces can be optimized when \code{scale = TRUE}

\code{scale.surfaces} can be used to specify which surfaces to apply kernel smoothing to during multisurface optimization. For example, \code{scale.surfaces = c(1, 0, 1)} will result in the first and third surfaces being optimized with a kernel smoothing function, while the second surface will not be scaled. The order of surfaces will match either the order of the raster stack, or alphabetical order when reading in from a directory.

\code{min.scale} defaults to a minimum of 0.01. During optimization, whenever the scaling factor (sigma) is less than 0.5, ResistanceGA will not apply scaling. In this way, it is possible for a surface to not be scaled.

The Default for \code{k.value} is 2, which sets k equal to the number of parameters optimized, plus 1 for the intercept term. Prior to version 3.0-0, \code{k.value} could not be specified by the user and followed setting 2, such that k was equal to the number of parameters optimized plus the intercept term.

\code{cont.shape} can take values of "Increase", "Decrease", or "Peaked". If you believe a resistance surface is related to your response in a particular way, specifying this here may decrease the time to optimization. \code{cont.shape} is used to generate an initial set of parameter values to test during optimization. If specified, a greater proportion of the starting values will include your believed relationship. If unspecified (the Default), a completely random set of starting values will be generated.

If it is desired that only certain transformations be assessed for continuous surfaces, then this can be specified using \code{select.trans}. By default, only monomolecular transformations will be assessed for continuous surfaces unless otherwise specified. Specific transformations can be specified by providing a vector of values (e.g., \code{c(1,3,5)}), with values corresponding to the equation numbers as detailed in \code{\link[ResistanceGA]{Resistance.tran}}. If multiple rasters are to be optimized from the same directory, then a list of transformations must be provided in the order that the raster surfaces will be assessed. For example:\cr
\code{select.trans = list("M", "A", "R", c(5,6))}\cr
will result in surface one only being optimized with Monomolecular transformations, surface two with all transformations, surface three with only Ricker transformations, and surface four with Reverse Ricker and Reverse Monomolecular only. If a categorical surface is among the rasters to be optimized, it is necessary to specify \code{NA} to accommodate this.

Setting \code{gaisl = TRUE} has the potential greatly reduce the optimization run time, potentially with greater accuracy. This is a distributed multiple-population GA, where the population is partitioned into several subpopulations and assigned to separated islands. Independent GAs are executed in each island, and only occasionally sparse exchanges of individuals are performed among the islands. 

It is recommended to first run GA optimization with the default settings
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
