% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdistance.R
\name{gdist.prep}
\alias{gdist.prep}
\title{Prepare data for optimization using \code{gdistance}}
\usage{
gdist.prep(n.Pops, response, samples, transitionFunction, directions, longlat, method)
}
\arguments{
\item{n.Pops}{The number of populations that are being assessed}

\item{response}{Vector of pairwise genetic distances (lower half of pairwise matrix).}

\item{samples}{Either provide the path to a .txt file containing the xy coordinates, or provide a matrix with x values in column 1 and y values in column 2. Alternatively, you can provide a \code{\link[sp]{SpatialPoints}} object}

\item{transitionFunction}{The function to calculate the gdistance TransitionLayer object. See \code{\link[gdistance]{transition}}. Default = function(x) 1/mean(x)}

\item{directions}{Directions in which cells are connected (4, 8, 16, or other). Default = 8}

\item{longlat}{Logical. If true, a \code{\link[gdistance]{geoCorrection}} ill be applied to the transition  matrix. Defaault = FALSE}

\item{method}{Specify whether pairwise distance should be calulated using the \code{\link[gdistance]{costDistance}} or \code{\link[gdistance]{commuteDistance}} (Default) functions. \code{\link[gdistance]{costDistance}} calculates least cost path distance, \code{\link[gdistance]{commuteDistance}} is equivalent (i.e. nearly perfectly correlated with) resistance distance calculated by CIRCUITSCAPE.}
}
\value{
An R object that is a required input into optimization functions
}
\description{
Creates a necessary input for optimizing resistance surfaces based on pairwise cost distances, implemented using the \code{gdistance} library
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
