% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gdist.prep}
\alias{gdist.prep}
\title{Prepare data for optimization using \code{gdistance}}
\usage{
gdist.prep(n.POPS, response, samples, transitionFunction, directions, longlat)
}
\arguments{
\item{n.POPS}{The number of populations that are being assessed}

\item{response}{Vector of pairwise genetic distances (lower half of pairwise matrix).}

\item{samples}{Either provide the path to a .txt file containing the xy coordinates, or provide a matrix with x values in column 1 and y values in column 2. Alternatively, you can provide a \code{\link[sp]{SpatialPoints}} object}

\item{transitionFunction}{The function to calculate the gdistance TransitionLayer object. See \code{\link[gdistance]{transition}}. Default = function(x) 1/mean(x)}

\item{directions}{Directions in which cells are connected (4, 8, 16, or other). Default = 8}

\item{longlat}{Logical. If true, a \code{\link[gdistance]{geoCorrection}} ill be applied to the transition  matrix. Defaault = FALSE}
}
\value{
An R object that is a required input into optimization functions
}
\description{
Creates a necessary input for optimizing resistance surfaces based on pairwise cost distances, implemented using the \code{gdistance} library
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

