% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdistance.R
\name{gdist.prep}
\alias{gdist.prep}
\title{Prepare data for optimization using \code{gdistance}}
\usage{
gdist.prep(n.Pops, 
                  response = NULL,
                  samples,
                  covariates = NULL,
                  formula = NULL,
                  transitionFunction = function(x)  1 / mean(x),
                  directions = 8,
                  longlat = FALSE,
                  method = 'commuteDistance',
                  min.max_dist = NULL,
                  keep = NULL)
}
\arguments{
\item{n.Pops}{The number of populations that are being assessed}

\item{response}{Vector of pairwise genetic distances (lower half of pairwise matrix).}

\item{samples}{Either provide the path to a .txt file containing the xy coordinates, or provide a matrix with x values in column 1 and y values in column 2. Alternatively, you can provide a \code{\link[sp]{SpatialPoints}} object}

\item{covariates}{Data frame of additional covariates that you want included in the MLPE model during opitmization.}

\item{formula}{If covariates are included in the model, specify the R formula for the fixed effects portion of the MLPE model.}

\item{transitionFunction}{The function to calculate the gdistance TransitionLayer object. See \code{\link[gdistance]{transition}}. Default = function(x) 1/mean(x)}

\item{directions}{Directions in which cells are connected (4, 8, 16, or other). Default = 8}

\item{longlat}{Logical. If true, a \code{\link[gdistance]{geoCorrection}} will be applied to the transition  matrix. Default = FALSE}

\item{method}{Specify whether pairwise distance should be calulated using the \code{\link[gdistance]{costDistance}} or \code{\link[gdistance]{commuteDistance}} (Default) functions. \code{\link[gdistance]{costDistance}} calculates least cost path distance, \code{\link[gdistance]{commuteDistance}} is equivalent (i.e. nearly perfectly correlated with) resistance distance calculated by CIRCUITSCAPE.}

\item{min.max_dist}{NOT YET SUPPORTED. Optional. Specify the minimum and maximum distance at which pairwise comparisons will be made(e.g., c(1, 50)). Euclidean distances below and above the minumum and maximum values will be omitted from the analysis. This has potential to reduce analysis time, but also reduces the number of pairwise comparisons.}

\item{keep}{NOT YET SUPPORTED. An optional vector equal to length \code{response} (i.e. all pairwise observations), with 1 indicating to keep the observation, and 0 to drop the observation. This can be used in conjunction with, or in place of \code{min.max_dist} to select which observations to include in analyses.}
}
\value{
An R object that is a required input into optimization functions
}
\description{
Creates a necessary input for optimizing resistance surfaces based on pairwise cost distances, implemented using the \code{gdistance} library
}
\details{
When specifying a formula, provide it as: \code{response ~ covariate}.
the formula \code{response} will use the vector of values specified for the \code{response} parameter. Make sure that covariate names match variable names provided in \code{covariates}
}
\examples{
 
## Not run:
## *** TO BE COMPLETED *** ##

## End (Not run)
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
