% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA_Functions.R
\name{lower}
\alias{lower}
\title{Make a vector of the lower half of a square distance matrix}
\usage{
lower(matrix)
}
\arguments{
\item{matrix}{Square distance matrix with no row names.}
}
\value{
A vector of the lower half of the matrix
}
\description{
This function will prepare and compile objects and commands for optimization with the GA package
}
\details{
This is a convenience function to obtain the lower half of a matrix, which is required as input for several other functions
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

