% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MLPE.lmm}
\alias{MLPE.lmm}
\title{Run maximum likelihood population effects mixed effects model (MLPE)}
\usage{
MLPE.lmm(cs.resistance, pairwise.genetic)
}
\arguments{
\item{cs.resistance}{Pairwise resistance distance matrix (resistances.out) from CS results}

\item{genetic.dist}{Lower half of pairwise genetic distance matrix}

\item{REML}{Logical. If TRUE, mixed effects model will be fit using restricted maximum likelihood. Default = TRUE.}
}
\value{
A glmer object from the fitted model
}
\description{
Runs MLPE as detailed by Clarke et al. (2002). This function will run the model and return glmer object
}
\references{
Clarke, R. T., P. Rothery, and A. F. Raybould. 2002. Confidence limits for regression relationships between distance matrices: Estimating gene flow with distance. Journal of Agricultural, Biological, and Environmental Statistics 7:361-372.
}

