% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ResistanceGA_Functions.R
\name{MLPE.lmm}
\alias{MLPE.lmm}
\title{Run maximum likelihood population effects mixed effects model (MLPE)}
\usage{
MLPE.lmm(resistance, pairwise.genetic, REML, ID, ZZ)
}
\arguments{
\item{resistance}{Path to pairwise resistance distance matrix (resistances.out) from CS results. Alternatively, provide the pairwise resistances created from optimizing with `gdistance` (result of Run_gdistance).}

\item{pairwise.genetic}{Lower half of pairwise genetic distance matrix}

\item{REML}{Logical. If TRUE, mixed effects model will be fit using restricted maximum likelihood. Default = FALSE}

\item{ID}{The to_from ID list for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of CS.prep or gdist.prep (Default = NULL)}

\item{ZZ}{The sparse matrix object for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of CS.prep or gdist.prep (Default = NULL)}
}
\value{
A lmer object from the fitted model
}
\description{
Runs MLPE as detailed by Clarke et al. (2002). This function will run the model and return a lmer object
}
\details{
An AIC value will only be returned if \code{REML = FALSE}
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
\references{
Clarke, R. T., P. Rothery, and A. F. Raybould. 2002. Confidence limits for regression relationships between distance matrices: Estimating gene flow with distance. Journal of Agricultural, Biological, and Environmental Statistics 7:361-372.
}

