% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rga_mlpe.R
\name{mlpe_rga}
\alias{mlpe_rga}
\title{Run maximum likelihood population effects mixed effects model (MLPE)}
\usage{
mlpe_rga(formula, data, REML = FALSE, ZZ = NULL)
}
\arguments{
\item{formula}{\code{lme4} style mixed effects model equation.}

\item{data}{data frame containing vectors of values from the lower half of genetic/resistance distance matrices}

\item{REML}{Logical. If TRUE, mixed effects model will be fit using restricted maximum likelihood. Default = FALSE}

\item{ZZ}{The sparse matrix object for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of CS.prep or gdist.prep (Default = NULL)}
}
\value{
A lmer object from the fitted model
}
\description{
Runs MLPE as detailed by Clarke et al. (2002). This is a general function for flexibly fitting MLPE models using the standard \code{lme4} formula interface
}
\details{
An AIC value will only be returned if \code{REML = FALSE}. The random effect must be the population vector \code{pop1} generated from the function \code{\link[ResistanceGA]{To.From.ID}}.
}
\examples{
 
# Create square 'distance' matrices
y <- matrix(rnorm(16), 4)
x <- matrix(rnorm(16), 4)

# Create to-from object (4 populations sampled)
id <- To.From.ID(4)

# Create data frame
df <- data.frame(y = lower(y),
                 x = lower(x),
                 pop = id$pop1)

# Fit MLPE model
out <- mlpe_rga(formula = y ~ x + (1 | pop),
                data = df)
}
\references{
Clarke, R. T., P. Rothery, and A. F. Raybould. 2002. Confidence limits for regression relationships between distance matrices: Estimating gene flow with distance. Journal of Agricultural, Biological, and Environmental Statistics 7:361-372.
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
