% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootsrap_function.R
\name{Resist.boot}
\alias{Resist.boot}
\title{Run bootstrap on optimized resistance surfaces}
\usage{
Resist.boot(mod.names, dist.mat, n.parameters, sample.prop, iters, obs, genetic.mat)
}
\arguments{
\item{mod.names}{A vector of the model names to be assessed}

\item{dist.mat}{A list containing all distance matrices from optimized resistance surfaces}

\item{n.parameters}{A vector the length of mod.names, specifying the number of parameters in each model}

\item{sample.prop}{Proportion of observations to be sampled each iteration (Default = 0.75)}

\item{iters}{DNumber of bootstrap iterations to be conducted}

\item{obs}{Total number of observations (populations or individuals) in your original analysis}

\item{genetic.mat}{Genetic distance matrix without row or column names.}
}
\value{
A data frame reporting the average model weight, average rank, number of time a model was the top model in teh set, and the frequency a model was best.
}
\description{
This is a 'pseudo' bootstrap procedure that will subsample a specified proportion of the sample locations/individuals, and will refit the MLPE model using the previously optimized resistance surface.
}
\details{
This is a 'pseudo'bootstrap procedure that subsamples distance and genetic matrices, refits the MLPE model for each surface, and assesses the model AICc. AICc is calculated based on the number of parameters specified.
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
