% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resistance_transform.R
\name{Resistance.tran}
\alias{Resistance.tran}
\title{Apply transformation to continuous resistance surface}
\usage{
Resistance.tran(transformation, shape, max, scale, r, out)
}
\arguments{
\item{transformation}{Transformation equation to apply. Can be provided as the name of the transformation or its numeric equivalent (see details)}

\item{shape}{Value of the shape parameter}

\item{max}{Value of the maximum value parameter}

\item{scale}{The standard deviation, in number of raster cells, to use when applying Gaussian kernel smoothing. This is the `sigma` parameter in the `kernel2dsmooth` function. (Default = NULL)}

\item{r}{Resistance surface to be transformed. Can be supplied as full path to .asc file or as a raster object}

\item{out}{Directory to write transformed .asc file. Default is NULL, and will not export .asc file}
}
\value{
R raster object
}
\description{
Apply one the eight resistance transformations to a continuous resistance surface
}
\details{
Valid arguements for \code{transformation} are:\cr
\tabular{ll}{
   \tab 1 = "Inverse-Reverse Monomolecular"\cr
   \tab 2 = "Inverse-Reverse Ricker"\cr
   \tab 3 = "Monomolecular"\cr
   \tab 4 = "Ricker"\cr
   \tab 5 = "Reverse Monomolecular"\cr
   \tab 6 = "Reverse Ricker"\cr
   \tab 7 = "Inverse Monomolecular"\cr
   \tab 8 = "Inverse Ricker"\cr
   \tab 9 = "Distance"\cr
   }

The Distance transformation sets all values equal to one. Because of the flexibility of the Ricker function to take a monomolecular shape (try \code{Plot.trans(PARM=c(10,100), Resistance=c(1,10), transformation="Ricker")} to see this), whenever a shape parameter >6 is selected in combination with a Ricker family transformation, the transformation reverts to a Distance transformation. In general, it seems that using a combination of intermediate Ricker and Monomolecular transformations provides the best, most flexible coverasge of parameter space.
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

