% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{resistance_surfaces}
\alias{resistance_surfaces}
\title{Simulated resistance surfaces}
\format{Raster stack object of length 4}
\usage{
data(resistance_surfaces)
}
\description{
A raster stack containing four raster surfaces
}
\details{
\itemize{
   \item continuous1. Continuous resistance surface
   \item continuous2. Continuous resistance surface
   \item categorical. 3-class categorical resistance surface
   \item feature. 2-class categorical (feature) resistance surface
   }
}
\keyword{datasets}

