% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{resistance_surfaces}
\alias{resistance_surfaces}
\title{Simulated resistance surfaces}
\format{RasterStack object of length 3}
\usage{
data(resistance_surfaces)
}
\description{
A raster stack containing three raster surfaces
}
\details{
\itemize{
   \item continuous. Continuous resistance surface
   \item categorical. 3-class categorical resistance surface
   \item feature. 2-class categorical (feature) resistance surface
   }
}
\keyword{datasets}

