% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{resistance_surfaces}
\alias{resistance_surfaces}
\title{Data file. Simulated resistance surfaces}
\format{RasterStack object of length 3}
\usage{
data(resistance_surfaces)
}
\description{
A raster stack containing three raster surfaces
}
\details{
\itemize{#'    
   \item categorical. 3-class categorical resistance surface
   \item continuous. Continuous resistance surface
   \item feature. 2-class categorical (feature) resistance surface
   }
}
\keyword{datasets}
