% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_julia.R
\name{Run_CS.jl}
\alias{Run_CS.jl}
\title{Run Julia version of CIRCUITSCAPE from R}
\usage{
Run_CS.jl(jl.inputs = NULL,
r,
CurrentMap = FALSE,
full.mat = FALSE,
EXPORT.dir = NULL,
output = "matrix",
CS_Point.File = NULL,
pairs_to_include = NULL,
pop2ind = NULL,
parallel = FALSE, 
cores = NULL,
cholmod = TRUE,
precision = FALSE, 
JULIA_HOME = NULL,
Julia_link = NULL,
rm.files = TRUE,
scratch = NULL,
is_resistance = TRUE)
}
\arguments{
\item{jl.inputs}{Object created from running \code{\link[ResistanceGA]{jl.prep}} function}

\item{r}{Accepts two types of inputs. Provide either the path to the resistance surface file (.asc) or specify an R RasterLayer object}

\item{CurrentMap}{Logical. If TRUE, the cumulative current resistance map will be generated during the CS run (Default = FALSE)}

\item{full.mat}{Logical (Default = FALSE). If TRUE, the full distance matrix will be generated as an R object, rather than just the lower half of the distance matrix.}

\item{EXPORT.dir}{Directory where CS results will be written. Only specify if Circuitscape current map outputs are requested. It is critical that there are NO SPACES in the directory, as this may cause the function to fail.}

\item{output}{Specifiy either "matrix" or "raster". "matrix" will return the lower half of the pairwise resistance matrix (default), while "raster" will return a \code{raster} object of the cumulative current map. The raster map can only be returned if \code{CurrentMap=TRUE}}

\item{CS_Point.File}{Provide a \code{\link[sp]{SpatialPoints}} object containing sample locations. Alternatively, specify the path to the Circuitscape formatted point file. See Circuitscape documentation for help. Only necessary to specify if \code{jl.inputs} are not specified.}

\item{pairs_to_include}{Default is NULL. If you wish to use the advanced CIRCUITSCAPE setting mode to include or exclude certain pairs of sample locations, provide a vector consisting of 1 (keep) or 0 (drop) for each pairwise observation (see example). This is an option if you do not want to assess all pairiwse observations. Only necessary to specify if \code{jl.inputs} are not specified.}

\item{pop2ind}{Default is NULL. A vector with length equivalent to the number of populations sapled. Each value of the vector indicates the number of samples from each respective population.}

\item{parallel}{(Logical; Default = FALSE) Do you want to run CIRCUITSCAPE in parallel? Only necessary to specify if \code{jl.inputs} are not specified.}

\item{cores}{If `parallel = TRUE`, how many cores should be used for parallel processing? Only necessary to specify if \code{jl.inputs} are not specified.}

\item{cholmod}{(Logical; Default = TRUE). Should the cholmod solver be used? See details. Only necessary to specify if \code{jl.inputs} are not specified.}

\item{precision}{(Logical; Default = FALSE). Should experimental single precision method be used? See details. Only necessary to specify if \code{jl.inputs} are not specified.}

\item{JULIA_HOME}{Path to the folder containing the Julia binary (See Details). Only necessary to specify if \code{jl.inputs} are not specified.}

\item{Julia_link}{Specify whether R should connect to Julia using the 'JuliaCall' package or the 'XRJulia' package. Will Default to using 'JuliaCall' if not specified here or in \code{jl.inputs}}

\item{rm.files}{Should all temporary files be removed after Julia run (Default = TRUE). Only necessary to specify if \code{jl.inputs} are not specified.}

\item{scratch}{Scratch directory for use if write access is limited. Must be specified if raster results are desired for outputs.}

\item{is_resistance}{Default = TRUE. Is the landscape represented as a resistance (TRUE) or conductance (FALSE) surface?}
}
\value{
Vector of CIRCUITSCAPE resistance distances (lower half of resistance matrix) OR a full square distance matrix if `full.mat` = TRUE. Alternatively, a raster object of the cumulative current map can be returned when \code{CurrentMap = TRUE} and \code{output = "raster"}. The `full.mat` cannot be requested if only select pairs are being analyzed.
}
\description{
Execute julia CS from R
}
\examples{
## To do

 
## Not run:
## *** TO BE COMPLETED *** ##

## End (Not run)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
