% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_julia.R
\name{Run_CS.jl}
\alias{Run_CS.jl}
\title{Run Julia version of CIRCUITSCAPE from R}
\usage{
Run_CS.jl(jl.inputs = NULL,
r,
CurrentMap = FALSE,
full.mat = FALSE,
EXPORT.dir = NULL,
output = "matrix",
CS_Point.File = NULL,
JULIA_HOME = NULL,
Julia_link = NULL,
rm.files = TRUE,
scratch = NULL)
}
\arguments{
\item{jl.inputs}{Object created from running \code{\link[ResistanceGA]{jl.prep}} function}

\item{r}{Accepts two types of inputs. Provide either the path to the resistance surface file (.asc) or specify an R RasterLayer object}

\item{CurrentMap}{Logical. If TRUE, the cumulative current resistance map will be generated during the CS run (Default = FALSE)}

\item{full.mat}{Logical (Default = FALSE). If TRUE, the full distance matrix will be generated as an R object, rather than just the lower half of the distance matrix.}

\item{EXPORT.dir}{Directory where CS results will be written. Only specify if Circuitscape current map outputs are requested. It is critical that there are NO SPACES in the directory, as this may cause the function to fail.}

\item{output}{Specifiy either "matrix" or "raster". "matrix" will return the lower half of the pairwise resistance matrix (default), while "raster" will return a \code{raster} object of the cumulative current map. The raster map can only be returned if \code{CurrentMap=TRUE}}

\item{CS_Point.File}{Provide a \code{\link[sp]{SpatialPoints}} object containing sample locations. Alternatively, specify the path to the Circuitscape formatted point file. See Circuitscape documentation for help. Only necessary to specify if \code{jl.inputs} are not specified.}

\item{JULIA_HOME}{Path to the folder containing the Julia binary (See Details). Only necessary to specify if \code{jl.inputs} are not specified.
#' @param Julia_link Specify whether R should connect to Julia using the 'JuliaCall' package or the 'XRJulia' package. Will Default to using 'JuliaCall' if not specified here or in \code{jl.inputs}}

\item{rm.files}{Should all temporary files be removed after Julia run (Default = TRUE)}

\item{scratch}{Scratch directory for use if write access is limited. Must be specified if raster results are desired for outputs.}
}
\value{
Vector of CIRCUITSCAPE resistance distances (lower half of resistance matrix) OR a full square distance matrix if `full.mat` = TRUE. Alternatively, a raster object of the cumulative current map can be returned when \code{CurrentMap = TRUE} and \code{output = "raster"}.
}
\description{
Execute julia CS from R
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
