% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_gdistance.R
\name{Run_gdistance}
\alias{Run_gdistance}
\title{Get cost distance using gdistance}
\usage{
Run_gdistance(gdist.inputs, r)
}
\arguments{
\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{CS.prep}} function}

\item{r}{Accepts two types of inputs. Provide either the path to the resistance surface file (.asc) or specify an R RasterLayer object}
}
\value{
A costDistance matrix object from gdistance
}
\description{
Execute gdistance
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
