% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_gdistance.R
\name{Run_gdistance}
\alias{Run_gdistance}
\title{Get cost distance using gdistance
Execute gdistance}
\usage{
Run_gdistance(gdist.inputs, r, scl)
}
\arguments{
\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA]{gdist.prep}} function}

\item{r}{Accepts two types of inputs. Provide either the path to the resistance surface file (.asc) or specify an R RasterLayer object}

\item{scl}{scale the correction values (default is TRUE). No scaling should be done if the user wants to obtain absolute distance values as output. See \code{\link[gdistance]{geoCorrection}} for details}
}
\value{
A costDistance matrix object from gdistance
}
\description{
Get cost distance using gdistance
Execute gdistance
}
\examples{
 
## Not run:
## *** TO BE COMPLETED *** ##

## End (Not run)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
