% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{samples}
\alias{samples}
\title{Example sample location file to run CIRCUITSCAPE}
\format{A 25 x 3 data frame}
\usage{
data(samples)
}
\description{
Sample file to be used with examples in the vignette
}
\details{
A data frame that can be saved as a .txt file for running examples in the vignette
}
\keyword{datasets}

