% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from.R
\name{To.From.ID}
\alias{To.From.ID}
\title{Convenience function to make to-from object needed for specifying MLPE random effects}
\usage{
To.From.ID(POPS)
}
\arguments{
\item{POPS}{Provide an integer value representing the number of individuals or populations that were sampled}
}
\value{
A data frame with two columns
}
\description{
Convenience function to make to-from object needed for specifying MLPE random effects
}
\details{
This object indicates the population pairs that distance values were calculated between. Note: Distance values must be taken from the lower half of a distance matrix. You can use \code{\link[ResistanceGA]{lower}} to obtain these values. When specifying a MLPE model using \code{\link[ResistanceGA]{mlpe_rga}}, \code{pop1} from \code{To.From.ID} must be added to the data frame containing model data, and must be specified as the random effect.
}
\examples{
To.From.ID(3)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}
